구조 1차 수정 progress > utils 통합
