--- atomicwriter.egg-info/SOURCES.txt
@@ -1,15 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-setup.py
-atomicwriter/__init__.py
-atomicwriter/core.py
-atomicwriter/plugins.py
-atomicwriter/utils.py
-atomicwriter.egg-info/PKG-INFO
-atomicwriter.egg-info/SOURCES.txt
-atomicwriter.egg-info/dependency_links.txt
-atomicwriter.egg-info/requires.txt
-atomicwriter.egg-info/top_level.txt
-tests/test_core.py
-tests/test_plugins.py
