--- demo.csv
@@ -1,4 +1,0 @@
-,a
-0,1
-1,2
-2,3
