--- demo.parquet._SUCCESS
@@ -1,1 +1,0 @@
-OK
