+++ src/atomicwriter/utils.py
@@ -1,7 +1,21 @@
-import time
+import logging
 import os
-import threading
 
+def setup_logger(name="atomicwriter"):
+    logger = logging.getLogger(name)
+    if not logger.handlers:
+        handler = logging.StreamHandler()
+        formatter = logging.Formatter("[%(levelname)s] %(message)s")
+        handler.setFormatter(formatter)
+        logger.addHandler(handler)
+    logger.setLevel(logging.INFO)
+    return logger
+
+def check_file_exists(path):
+    return os.path.exists(path)
+
+import time
+import threading
 
 class ProgressBar:
     """
