--- tests/output_data/demo_pandas.csv
@@ -1,4 +1,0 @@
-,a
-0,1
-1,2
-2,3
