--- tests/output_data/demo_pandas.csv._SUCCESS
@@ -1,1 +1,0 @@
-OK
