--- tests/output_data/demo_pandas.html._SUCCESS
@@ -1,1 +1,0 @@
-OK
