--- tests/output_data/demo_pandas.json
@@ -1,1 +1,0 @@
-{"a":{"0":1,"1":2,"2":3}}
