--- tests/output_data/demo_pandas.json._SUCCESS
@@ -1,1 +1,0 @@
-OK
