--- tests/output_data/demo_pandas.parquet._SUCCESS
@@ -1,1 +1,0 @@
-OK
