--- tests/output_data/demo_pandas.pkl._SUCCESS
@@ -1,1 +1,0 @@
-OK
