--- tests/output_data/demo_polars.avro._SUCCESS
@@ -1,1 +1,0 @@
-OK
