--- tests/output_data/demo_polars.csv
@@ -1,4 +1,0 @@
-a
-1
-2
-3
