--- tests/output_data/demo_polars.csv._SUCCESS
@@ -1,1 +1,0 @@
-OK
