--- tests/output_data/demo_polars.ipc._SUCCESS
@@ -1,1 +1,0 @@
-OK
