--- tests/output_data/demo_polars.json
@@ -1,1 +1,0 @@
-[{"a":1},{"a":2},{"a":3}]
