--- tests/output_data/demo_polars.json._SUCCESS
@@ -1,1 +1,0 @@
-OK
