--- tests/output_data/demo_polars.parquet._SUCCESS
@@ -1,1 +1,0 @@
-OK
